% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{capitan_log}
\alias{capitan_log}
\title{Append the entry in the main capitan's log.}
\usage{
capitan_log(stardate = as.character(Sys.time()), description = "",
  version = "1.0", number = 1, folder = "./logs/")
}
\arguments{
\item{stardate}{is the experiment execution date. Default: as.character(Sys.time())}

\item{description}{containing the main information of you experiment. E.g: Testing the max_date new feature}

\item{version}{a character with the main version of your project. E.g: "1.3" == v1.3 or just "1"}

\item{number}{of the experiment. E.g: 4 == "v1.3.4"}

\item{folder}{as character to store the logs}
}
\description{
"Private method" used to create and append the main capitans log containing the:
Complete version, Stardate and Description
}
\examples{
capitan_log(
    stardate = as.character(Sys.time()),
    description = "First experiment using XGBoost witout categorical features",
    tag = 'ml', version='1.0', number = 1
)

}
