% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{create_folders}
\alias{create_folders}
\title{Create the folder recursivelly, if exists}
\usage{
create_folders(folder)
}
\arguments{
\item{folder}{name as character}
}
\description{
Create the folder recursivelly, if exists
}
