% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{log_experiment}
\alias{log_experiment}
\title{Create an experiment entry log}
\usage{
log_experiment(..., description = "", tag = "", version = "1.0",
  number = 1, folder = "./logs/")
}
\arguments{
\item{...}{all the objects and information you want to log. We recommend a ("title", value, "") format.}

\item{description}{containing the main information of you experiment. E.g: Testing the max_date new feature}

\item{tag}{an optional tag to separate and organize your logs.}

\item{version}{a character with the main version of your project. E.g: "1.3" == v1.3 or just "1"}

\item{number}{of the experiment. E.g: 4 == "v1.3.4"}

\item{folder}{as character to store the logs}
}
\description{
Create a new text file in the selected {folder} with all the experiment
information passed refered to an experiment, the context and the result.
}
\details{
All the information should be passed a ... parameters.
Empty characters are considered new lines.

Supports: characters, numbers, data.frames, lists, vectors, etc.
}
\examples{
log_experiment(
    # Entry log information:
    description = "First experiment using XGBoost witout categorical features",
    tag = 'ml', version='1.0', number = 1,
    # Log all the needed information:
    "Final AUC:", 0.789, ""
)

}
\seealso{
\code{\link{capitan_log}}
}
